pyproject.toml
[project]
name = "referral-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.4",
    "django-apscheduler>=0.7.0",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
    "gunicorn>=23.0.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
]
